Main with basic no zero condition
This file contains the condition that none of the points in the inner aperture can be equal to zero. E.g. if we mask a galaxy non completely, this prevents the residual from being calculated as an additional galaxy. However it doesn't work with the background shell, so technically the background can be lowered.